# apps/api/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API HTTP do PetControl para autenticação, módulos de tenant e agendamentos.",
        "title": "PetControl API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/company-users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lists users linked to the authenticated tenant company.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "company_users"
                ],
                "summary": "List company users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.CompanyUserListResponseDoc"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handler.APIErrorResponseDoc"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.APIErrorResponseDoc"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Links an existing user to the authenticated tenant company. Requires company owner access.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "company_users"
                ],
                "summary": "Create company user link",
                "parameters": [
                    {
                        "description": "Company user payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CompanyUserCreateRequestDoc"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.CompanyUserItemResponseDoc"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handler.APIErrorResponseDoc"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/handler.APIErrorResponseDoc"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.APIErrorResponseDoc"
                        }
                    }
                }
            }
        },
        "/company-users/{user_id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deactivates a user link for the authenticated tenant company. Requires company owner access.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "company_users"
                ],
                "summary": "Deactivate company user link",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handler.APIErrorResponseDoc"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.APIErrorResponseDoc"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/handler.APIErrorResponseDoc"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.APIErrorResponseDoc"
                        }
                    }
                }
            }
        },
        "/company-users/{user_id}/permissions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lists the configurable tenant settings permissions for a user linked to the authenticated company. Admin only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "company_users"
                ],
                "summary": "List company user manageable permissions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.CompanyUserPermissionsResponseDoc"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handler.APIErrorResponseDoc"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.APIErrorResponseDoc"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/handler.APIErrorResponseDoc"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.APIErrorResponseDoc"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the configurable tenant settings permissions for a user linked to the authenticated company. Admin only.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "company_users"
                ],
                "summary": "Update company user manageable permissions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Permissions payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CompanyUserPermissionsUpdateRequestDoc"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.CompanyUserPermissionsResponseDoc"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handler.APIErrorResponseDoc"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.APIErrorResponseDoc"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/handler.APIErrorResponseDoc"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.APIErrorResponseDoc"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.APIErrorResponseDoc": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "correlation_id": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "handler.CompanyUserCreateRequestDoc": {
            "type": "object",
            "properties": {
                "is_owner": {
                    "type": "boolean",
                    "example": false
                },
                "user_id": {
                    "type": "string",
                    "example": "22222222-2222-2222-2222-222222222222"
                }
            }
        },
        "handler.CompanyUserDoc": {
            "type": "object",
            "properties": {
                "company_id": {
                    "type": "string",
                    "example": "11111111-1111-1111-1111-111111111111"
                },
                "full_name": {
                    "type": "string",
                    "example": "System PetControl"
                },
                "id": {
                    "type": "string",
                    "example": "dddddddd-dddd-dddd-dddd-dddddddddddd"
                },
                "image_url": {
                    "type": "string",
                    "example": "https://cdn.example.com/users/system.png"
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "is_owner": {
                    "type": "boolean",
                    "example": false
                },
                "joined_at": {
                    "type": "string",
                    "example": "2026-04-10T10:00:00Z"
                },
                "kind": {
                    "type": "string",
                    "example": "employee"
                },
                "left_at": {
                    "type": "string",
                    "example": "2026-04-11T11:00:00Z"
                },
                "role": {
                    "type": "string",
                    "example": "system"
                },
                "short_name": {
                    "type": "string",
                    "example": "System"
                },
                "user_id": {
                    "type": "string",
                    "example": "22222222-2222-2222-2222-222222222222"
                }
            }
        },
        "handler.CompanyUserItemResponseDoc": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/handler.CompanyUserDoc"
                }
            }
        },
        "handler.CompanyUserListResponseDoc": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.CompanyUserDoc"
                    }
                }
            }
        },
        "handler.CompanyUserPermissionDoc": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "company_settings:edit"
                },
                "default_roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "root",
                        "admin"
                    ]
                },
                "description": {
                    "type": "string",
                    "example": "Editar configurações gerais"
                },
                "granted_at": {
                    "type": "string",
                    "example": "2026-04-20T09:30:00Z"
                },
                "granted_by": {
                    "type": "string",
                    "example": "33333333-3333-3333-3333-333333333333"
                },
                "id": {
                    "type": "string",
                    "example": "44444444-4444-4444-4444-444444444444"
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "is_default_for_role": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "handler.CompanyUserPermissionsDoc": {
            "type": "object",
            "properties": {
                "company_id": {
                    "type": "string",
                    "example": "11111111-1111-1111-1111-111111111111"
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "is_owner": {
                    "type": "boolean",
                    "example": false
                },
                "kind": {
                    "type": "string",
                    "example": "employee"
                },
                "managed_by": {
                    "type": "string",
                    "example": "33333333-3333-3333-3333-333333333333"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.CompanyUserPermissionDoc"
                    }
                },
                "role": {
                    "type": "string",
                    "example": "system"
                },
                "scope": {
                    "type": "string",
                    "example": "tenant_settings"
                },
                "user_id": {
                    "type": "string",
                    "example": "22222222-2222-2222-2222-222222222222"
                }
            }
        },
        "handler.CompanyUserPermissionsResponseDoc": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/handler.CompanyUserPermissionsDoc"
                }
            }
        },
        "handler.CompanyUserPermissionsUpdateRequestDoc": {
            "type": "object",
            "properties": {
                "permission_codes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "plan_settings:edit",
                        "notification_settings:edit"
                    ]
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Use o formato: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}